+++ app/javascript/react/components/Map/Map.tsx
@@ -38,10 +38,10 @@ import useMobileDetection from "../../utils/useScreenSizeDetection";
 import { updateAll } from "../../store/thresholdSlice";
 import { MobileStreamShortInfo as StreamShortInfo } from "../../types/mobileStream";
 import { selectFixedStreamShortInfo } from "../../store/fixedStreamSelectors";
-import { SessionsListView } from "../SessionsListView/Test";
+import { SessionsListView } from "../SessionsListView/SessionsListView";
 import { SectionButton } from "../SectionButton/SectionButtonTest";
 import pinImage from "../../assets/icons/pinImage.svg";
-import { MobileSessionList } from "../SessionsListView/MobileSessionList/MobileSessionListTest";
+import { MobileSessionList } from "../SessionsListView/MobileSessionList/MobileSessionList";
 import { SessionList } from "../../types/sessionType";
 import { pubSub } from "../../utils/pubSubManager";
 import * as S from "./Map.style";
