+++ app/javascript/react/components/SessionsListView/MobileSessionList/MobileSessionList.style.tsx
