+++ app/javascript/react/components/SessionsListView/MobileSessionList/index.ts
@@ -1,1 +1,1 @@
-export * from "./MobileSessionListTest";
+export * from "./MobileSessionList";
