+++ app/javascript/react/components/SessionsListView/SessionsListTile/SessionListTile.style.tsx
