+++ app/javascript/react/components/SessionsListView/SessionsListTile/index.ts
@@ -1,1 +1,1 @@
-export * from "./SessionListTileT";
+export * from "./SessionListTile";
