+++ app/javascript/react/components/SessionsListView/SessionsListView.style.tsx
