+++ app/javascript/react/components/SessionsListView/index.ts
@@ -1,1 +1,1 @@
-export * from "./Test";
+export * from "./SessionsListView";
